StockBot/config.py
import os
from dotenv import load_dotenv

load_dotenv()

# Alpha Vantage API 키 설정
ALPHA_VANTAGE_API_KEY = os.getenv("ALPHA_VANTAGE_API_KEY")

# Google API 키 및 검색엔진 ID
GOOGLE_API_KEY = os.getenv("GOOGLE_API_KEY")
SEARCH_ENGINE_ID = os.getenv("SEARCH_ENGINE_ID")

# Hugging Face 키 설정
HUGGINGFACEHUB_API_TOKEN = os.getenv("HUGGINGFACEHUB_API_TOKEN")
