StockBot/requirements.txt
streamlit
gradio
yfinance
matplotlib
python-dotenv
pandas
requests
alpha_vantage
torch
langchain
langchain-community
langchain-huggingface
sentence-transformers
transformers
faiss-cpu
scikit-learn
langchain-ollama
mplfinance
